test RepositorioMascota

+++ src/main/java/com/tallerwebi/dominio/RepositorioMascota.java
@@ -1,7 +1,7 @@
 package com.tallerwebi.dominio;
 
 public interface RepositorioMascota {
-    Mascota buscarMascotaPorId(Integer idMascota);
+    Mascota buscarMascotaPorId(Long idMascota);
     void guardarMascota(Mascota mascota);
     void modificarMascota(Mascota mascota);
     void eliminarMascota(Mascota mascota);
